Fetch single order sync

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/store/WCOrderStore.kt
@@ -603,6 +603,19 @@ class WCOrderStore @Inject constructor(
         }
     }
 
+    suspend fun fetchSingleOrderSync(site: SiteModel, remoteOrderId: Long): WooResult<OrderEntity> {
+        return coroutineEngine.withDefaultContext(API, this, "fetchSingleOrderSync") {
+            val result = wcOrderRestClient.fetchSingleOrder(site, remoteOrderId)
+
+            return@withDefaultContext if (result.isError) {
+                WooResult(WooError(API_ERROR, SERVER_ERROR, result.error.message))
+            } else {
+                insertOrder(site.localId(), result.orderWithMeta)
+                WooResult(result.order)
+            }
+        }
+    }
+
     @Suppress("SpreadOperator")
     suspend fun fetchOrders(
         site: SiteModel,
@@ -1094,7 +1107,10 @@ class WCOrderStore @Inject constructor(
         return deleteOptions
     }
 
-    suspend fun fetchOrdersListFirstPage(listDescriptor: WCOrderListDescriptor): WooResult<List<OrderEntity>> {
+    suspend fun fetchOrdersListFirstPage(
+        listDescriptor: WCOrderListDescriptor,
+        deleteOldData: Boolean = false
+    ): WooResult<List<OrderEntity>> {
         val response = wcOrderRestClient.fetchOrdersListFirstPage(listDescriptor)
         return if (response.isError) {
             WooResult(WooError(API_ERROR, SERVER_ERROR, response.error.message))
@@ -1113,7 +1129,10 @@ class WCOrderStore @Inject constructor(
                 OrderSqlUtils.insertOrUpdateOrderSummaries(orderSummaries)
             }
 
-            ordersDaoDecorator.deleteOrdersForSite(listDescriptor.site.localId())
+            if (deleteOldData) {
+                ordersDaoDecorator.deleteOrdersForSite(listDescriptor.site.localId())
+            }
+
             @Suppress("SpreadOperator")
             insertOrder(listDescriptor.site.localId(), *result.toTypedArray())
             WooResult(orders)
